+++ service/pkg/server/options.go
@@ -2,6 +2,7 @@ package server
 
 import (
 	"github.com/casbin/casbin/v2/persist"
+	"github.com/opentdf/platform/service/pkg/config"
 	"github.com/opentdf/platform/service/pkg/serviceregistry"
 )
 
@@ -12,10 +13,12 @@ type StartConfig struct {
 	ConfigFile            string
 	WaitForShutdownSignal bool
 	PublicRoutes          []string
+	IPCReauthRoutes       []string
 	bultinPolicyOverride  string
 	extraCoreServices     []serviceregistry.IService
 	extraServices         []serviceregistry.IService
 	casbinAdapter         persist.Adapter
+	configLoaders         []config.Loader
 }
 
 // Deprecated: Use WithConfigKey
@@ -61,6 +64,15 @@ func WithPublicRoutes(routes []string) StartOptions {
 	}
 }
 
+// WithIPCReauthRoutes option sets the IPC reauthorization routes for the server.
+// It enables the server to reauthorize IPC routes and embed the token on the context.
+func WithIPCReauthRoutes(routes []string) StartOptions {
+	return func(c StartConfig) StartConfig {
+		c.IPCReauthRoutes = routes
+		return c
+	}
+}
+
 // WithAuthZPolicy option sets the default casbin policy to be used.
 // Example:
 //
@@ -98,6 +110,14 @@ func WithServices(services ...serviceregistry.IService) StartOptions {
 func WithCasbinAdapter(adapter persist.Adapter) StartOptions {
 	return func(c StartConfig) StartConfig {
 		c.casbinAdapter = adapter
+		return c
+	}
+}
+
+// WithAdditionalConfigLoader option adds an additional configuration loader to the server.
+func WithAdditionalConfigLoader(loader config.Loader) StartOptions {
+	return func(c StartConfig) StartConfig {
+		c.configLoaders = append(c.configLoaders, loader)
 		return c
 	}
 }