+++ service/policy/attributes/attributes_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func getValidator() *protovalidate.Validator {
+func getValidator() protovalidate.Validator {
 	v, err := protovalidate.New()
 	if err != nil {
 		panic(err)