+++ service/policy/namespaces/namespaces_test.go
@@ -19,7 +19,7 @@ const (
 	errExclusiveFields = "exclusive_fields"
 )
 
-func getValidator() *protovalidate.Validator {
+func getValidator() protovalidate.Validator {
 	v, err := protovalidate.New()
 	if err != nil {
 		panic(err)